Amelioration du jeu
ajout de multiples fonctionnalités, début de refonte graphique et ajout de classes
